feat: extract components and improve folder structure
